+++ src/modules/mime/mime.cpp
@@ -196,7 +196,7 @@ int MODULE_EXPORT GetStorageItem(INT_PTR* storage, int item_index, LPWIN32_FIND_
 	wcscpy_s(item_data->cFileName, MAX_PATH, name.c_str());
 	wcscpy_s(item_data->cAlternateFileName, 14, L"");
 	item_data->dwFileAttributes = FILE_ATTRIBUTE_NORMAL;
-	item_data->nFileSizeLow = entity->body().length();
+	item_data->nFileSizeLow = (DWORD) entity->body().length();
 
 	return GET_ITEM_OK;
 }
